feat: new function for moving chromeos devices
A lot of refactoring of older code and the addition of a new function
that will be a standalone response action instead of piecing together a
bunch of collective actions.
TBD: Test making the Customer ID a parameter for the API integration so
that we don't have to get it from a collection.