feat: Removing UI page in place of Config Settings (#15)
* update ui extension

renames device status to match admin console and adds tooltip to button

* manifest indent updates

uses foundry's standard for inconsistent indentation

* fix: update settings ui and readme to use path instead of OU ID

* feat: new function for moving chromeos devices

A lot of refactoring of older code and the addition of a new function
that will be a standalone response action instead of piecing together a
bunch of collective actions.
TBD: Test making the Customer ID a parameter for the API integration so
that we don't have to get it from a collection.

* added list orgunits and clean up

* adding dynamic listing of ou's

* adding the ability to hopefully list your orgunits

* attempt 1 to install

* this is as far as I go.. time to revert and rethink

* feat: new ui settings with load ou's initial deploy

* Loading works and functionality seems to be there

This is a good stopping point for now.

* Revert "this is as far as I go.. time to revert and rethink"

This reverts commit 6ece26e523ef3cd269617800d6d312189b498550.

* feat: finally got func working with api int changes

* chore: update .gitignore

* removing stuff that should be ignored

* fix: removed loading settings since it n/a

* feat: added new handler func to change device status

* minor change

* chore: clean up some minor stuff

* ci: fix release workflow