fix: update error message to be more generic
basically if we can't get device data, I don't think we need to spend
too much time/effort trying to track every permuatation that may go
wrong from an error code perspective. If it ain't working, it's most
likely that a) your credentials are wrong or b) the device doesn't exist
in the Google workspace you are targetting.